[liquidcheck] Change Measure.conten from int to double, since the json response may contain a double.

+++ bundles/org.openhab.binding.liquidcheck/src/main/java/org/openhab/binding/liquidcheck/internal/json/Measure.java
@@ -23,7 +23,7 @@ import org.eclipse.jdt.annotation.NonNullByDefault;
 @NonNullByDefault
 public class Measure {
     public double level = 0.0;
-    public int content = 0;
+    public double content = 0.0;
     public Raw raw = new Raw();
     public int age = 0;
 }
